[server] Move prebuilds out of EE
