+++ components/server/src/bitbucket-server/bitbucket-server-container-module.ts
@@ -16,6 +16,8 @@ import { BitbucketServerFileProvider } from "./bitbucket-server-file-provider";
 import { BitbucketServerRepositoryProvider } from "./bitbucket-server-repository-provider";
 import { BitbucketServerTokenHelper } from "./bitbucket-server-token-handler";
 import { BitbucketServerTokenValidator } from "./bitbucket-server-token-validator";
+import { RepositoryService } from "../repohost/repo-service";
+import { BitbucketServerService } from "../prebuilds/bitbucket-server-service";
 
 export const bitbucketServerContainerModule = new ContainerModule((bind, _unbind, _isBound, _rebind) => {
     bind(RepositoryHost).toSelf().inSingletonScope();
@@ -31,4 +33,5 @@ export const bitbucketServerContainerModule = new ContainerModule((bind, _unbind
     bind(BitbucketServerTokenHelper).toSelf().inSingletonScope();
     bind(BitbucketServerTokenValidator).toSelf().inSingletonScope();
     bind(IGitTokenValidator).toService(BitbucketServerTokenValidator);
+    bind(RepositoryService).to(BitbucketServerService).inSingletonScope();
 });
