+++ components/server/src/github/github-app-support.ts
@@ -8,7 +8,7 @@ import { ProviderRepository, User } from "@gitpod/gitpod-protocol";
 import { inject, injectable } from "inversify";
 import { GithubApp } from "../prebuilds/github-app";
 import { RequestError } from "@octokit/request-error";
-import { TokenProvider } from "../../../src/user/token-provider";
+import { TokenProvider } from "../user/token-provider";
 import { UserDB } from "@gitpod/gitpod-db/lib";
 import { Octokit, RestEndpointMethodTypes } from "@octokit/rest";
 import { log } from "@gitpod/gitpod-protocol/lib/util/logging";
