+++ components/server/src/prebuilds/bitbucket-app.ts
@@ -8,12 +8,12 @@ import * as express from "express";
 import { postConstruct, injectable, inject } from "inversify";
 import { ProjectDB, TeamDB, UserDB, WebhookEventDB } from "@gitpod/gitpod-db/lib";
 import { User, StartPrebuildResult, CommitContext, CommitInfo, Project, WebhookEvent } from "@gitpod/gitpod-protocol";
-import { PrebuildManager } from "../prebuilds/prebuild-manager";
+import { PrebuildManager } from "./prebuild-manager";
 import { TraceContext } from "@gitpod/gitpod-protocol/lib/util/tracing";
-import { TokenService } from "../../../src/user/token-service";
-import { ContextParser } from "../../../src/workspace/context-parser-service";
-import { HostContextProvider } from "../../../src/auth/host-context-provider";
-import { RepoURL } from "../../../src/repohost";
+import { TokenService } from "../user/token-service";
+import { ContextParser } from "../workspace/context-parser-service";
+import { HostContextProvider } from "../auth/host-context-provider";
+import { RepoURL } from "../repohost";
 
 @injectable()
 export class BitbucketApp {
