+++ components/server/src/prebuilds/gitlab-app.ts
@@ -8,14 +8,14 @@ import * as express from "express";
 import { postConstruct, injectable, inject } from "inversify";
 import { ProjectDB, TeamDB, UserDB, WebhookEventDB } from "@gitpod/gitpod-db/lib";
 import { Project, User, StartPrebuildResult, CommitContext, CommitInfo, WebhookEvent } from "@gitpod/gitpod-protocol";
-import { PrebuildManager } from "../prebuilds/prebuild-manager";
+import { PrebuildManager } from "./prebuild-manager";
 import { TraceContext } from "@gitpod/gitpod-protocol/lib/util/tracing";
-import { TokenService } from "../../../src/user/token-service";
-import { HostContextProvider } from "../../../src/auth/host-context-provider";
+import { TokenService } from "../user/token-service";
+import { HostContextProvider } from "../auth/host-context-provider";
 import { GitlabService } from "./gitlab-service";
 import { log } from "@gitpod/gitpod-protocol/lib/util/logging";
-import { ContextParser } from "../../../src/workspace/context-parser-service";
-import { RepoURL } from "../../../src/repohost";
+import { ContextParser } from "../workspace/context-parser-service";
+import { RepoURL } from "../repohost";
 
 @injectable()
 export class GitLabApp {
