+++ components/server/src/workspace/gitpod-server-impl.ts
@@ -141,16 +141,17 @@ import { WorkspaceFactory } from "./workspace-factory";
 import { WorkspaceStarter } from "./workspace-starter";
 import { HeadlessLogUrls } from "@gitpod/gitpod-protocol/lib/headless-workspace-log";
 import { HeadlessLogService, HeadlessLogEndpoint } from "./headless-log-service";
-import { InvalidGitpodYMLError } from "./config-provider";
+import { ConfigProvider, InvalidGitpodYMLError } from "./config-provider";
 import { ProjectsService } from "../projects/projects-service";
 import { LocalMessageBroker } from "../messaging/local-message-broker";
 import { IDEOptions } from "@gitpod/gitpod-protocol/lib/ide-protocol";
 import { PartialProject, OrganizationSettings } from "@gitpod/gitpod-protocol/lib/teams-projects-protocol";
-import { ClientMetadata } from "../websocket/websocket-connection-manager";
+import { ClientMetadata, traceClientMetadata } from "../websocket/websocket-connection-manager";
 import {
     AdditionalUserData,
     EnvVarWithValue,
     LinkedInProfile,
+    OpenPrebuildContext,
     ProjectEnvVar,
     UserFeatureSettings,
     WorkspaceTimeoutSetting,
@@ -192,6 +193,11 @@ import { isWorkspaceRegion, WorkspaceRegion } from "@gitpod/gitpod-protocol/lib/
 import { EnvVarService } from "./env-var-service";
 import { LinkedInService } from "../linkedin-service";
 import { SnapshotService, WaitForSnapshotOptions } from "./snapshot-service";
+import { IncrementalPrebuildsService } from "../prebuilds/incremental-prebuilds-service";
+import { PrebuildManager } from "../prebuilds/prebuild-manager";
+import { GitHubAppSupport } from "../github/github-app-support";
+import { GitLabAppSupport } from "../gitlab/gitlab-app-support";
+import { BitbucketAppSupport } from "../bitbucket/bitbucket-app-support";
 
 // shortcut
 export const traceWI = (ctx: TraceContext, wi: Omit<LogContext, "userId">) => TraceContext.setOWI(ctx, wi); // userId is already taken care of in WebsocketConnectionManager
@@ -219,6 +225,13 @@ export class GitpodServerImpl implements GitpodServerWithTracing, Disposable {
     @inject(InstallationAdminTelemetryDataProvider)
     protected readonly telemetryDataProvider: InstallationAdminTelemetryDataProvider;
 
+    @inject(GitHubAppSupport) protected readonly githubAppSupport: GitHubAppSupport;
+    @inject(GitLabAppSupport) protected readonly gitLabAppSupport: GitLabAppSupport;
+    @inject(BitbucketAppSupport) protected readonly bitbucketAppSupport: BitbucketAppSupport;
+
+    @inject(PrebuildManager) protected readonly prebuildManager: PrebuildManager;
+    @inject(IncrementalPrebuildsService) protected readonly incrementalPrebuildsService: IncrementalPrebuildsService;
+    @inject(ConfigProvider) protected readonly configProvider: ConfigProvider;
     @inject(WorkspaceStarter) protected readonly workspaceStarter: WorkspaceStarter;
     @inject(SnapshotService) protected readonly snapshotService: SnapshotService;
     @inject(WorkspaceManagerClientProvider)
@@ -300,6 +313,225 @@ export class GitpodServerImpl implements GitpodServerWithTracing, Disposable {
         log.debug({ userId: this.user?.id }, "initializeClient");
 
         this.listenForWorkspaceInstanceUpdates();
+
+        this.listenForPrebuildUpdates().catch((err) => log.error("error registering for prebuild updates", err));
+    }
+
+    protected async listenForPrebuildUpdates() {
+        // 'registering for prebuild updates for all projects this user has access to
+        const projects = await this.getAccessibleProjects();
+        for (const projectId of projects) {
+            this.disposables.push(
+                this.localMessageBroker.listenForPrebuildUpdates(
+                    projectId,
+                    (ctx: TraceContext, update: PrebuildWithStatus) =>
+                        TraceContext.withSpan(
+                            "forwardPrebuildUpdateToClient",
+                            (ctx) => {
+                                traceClientMetadata(ctx, this.clientMetadata);
+                                TraceContext.setJsonRPCMetadata(ctx, "onPrebuildUpdate");
+
+                                this.client?.onPrebuildUpdate(update);
+                            },
+                            ctx,
+                        ),
+                ),
+            );
+        }
+
+        // TODO(at) we need to keep the list of accessible project up to date
+    }
+
+    protected async getAccessibleProjects() {
+        if (!this.user) {
+            return [];
+        }
+
+        // update all project this user has access to
+        const allProjects: string[] = [];
+        const teams = await this.teamDB.findTeamsByUser(this.user.id);
+        for (const team of teams) {
+            allProjects.push(...(await this.projectsService.getTeamProjects(team.id)).map((p) => p.id));
+        }
+        allProjects.push(...(await this.projectsService.getUserProjects(this.user.id)).map((p) => p.id));
+        return allProjects;
+    }
+
+    protected async findPrebuiltWorkspace(
+        parentCtx: TraceContext,
+        user: User,
+        context: WorkspaceContext,
+        ignoreRunningPrebuild?: boolean,
+        allowUsingPreviousPrebuilds?: boolean,
+    ): Promise<WorkspaceCreationResult | PrebuiltWorkspaceContext | undefined> {
+        const ctx = TraceContext.childContext("findPrebuiltWorkspace", parentCtx);
+        try {
+            if (!(CommitContext.is(context) && context.repository.cloneUrl && context.revision)) {
+                return;
+            }
+
+            const commitSHAs = CommitContext.computeHash(context);
+
+            const logCtx: LogContext = { userId: user.id };
+            const cloneUrl = context.repository.cloneUrl;
+            let prebuiltWorkspace: PrebuiltWorkspace | undefined;
+            const logPayload = {
+                allowUsingPreviousPrebuilds,
+                ignoreRunningPrebuild,
+                cloneUrl,
+                commit: commitSHAs,
+                prebuiltWorkspace,
+            };
+            if (OpenPrebuildContext.is(context)) {
+                prebuiltWorkspace = await this.workspaceDb.trace(ctx).findPrebuildByID(context.openPrebuildID);
+                if (
+                    prebuiltWorkspace?.cloneURL !== cloneUrl &&
+                    (ignoreRunningPrebuild || prebuiltWorkspace?.state === "available")
+                ) {
+                    // prevent users from opening arbitrary prebuilds this way - they must match the clone URL so that the resource guards are correct.
+                    return;
+                }
+            } else {
+                log.debug(logCtx, "Looking for prebuilt workspace: ", logPayload);
+                prebuiltWorkspace = await this.workspaceDb
+                    .trace(ctx)
+                    .findPrebuiltWorkspaceByCommit(cloneUrl, commitSHAs);
+                if (!prebuiltWorkspace && allowUsingPreviousPrebuilds) {
+                    const { config } = await this.configProvider.fetchConfig({}, user, context);
+                    const history = await this.incrementalPrebuildsService.getCommitHistoryForContext(context, user);
+                    prebuiltWorkspace = await this.incrementalPrebuildsService.findGoodBaseForIncrementalBuild(
+                        context,
+                        config,
+                        history,
+                        user,
+                    );
+                }
+            }
+            if (!prebuiltWorkspace) {
+                return;
+            }
+
+            if (prebuiltWorkspace.state === "available") {
+                log.info(logCtx, `Found prebuilt workspace for ${cloneUrl}:${commitSHAs}`, logPayload);
+                const result: PrebuiltWorkspaceContext = {
+                    title: context.title,
+                    originalContext: context,
+                    prebuiltWorkspace,
+                };
+                return result;
+            } else if (prebuiltWorkspace.state === "queued") {
+                // waiting for a prebuild that has not even started yet, doesn't make sense.
+                // starting a workspace from git will be faster anyway
+                return;
+            } else if (prebuiltWorkspace.state === "building") {
+                if (ignoreRunningPrebuild) {
+                    // in force mode we ignore running prebuilds as we want to start a workspace as quickly as we can.
+                    return;
+                }
+
+                const workspaceID = prebuiltWorkspace.buildWorkspaceId;
+                const makeResult = (instanceID: string): WorkspaceCreationResult => {
+                    return <WorkspaceCreationResult>{
+                        runningWorkspacePrebuild: {
+                            prebuildID: prebuiltWorkspace!.id,
+                            workspaceID,
+                            instanceID,
+                            starting: "queued",
+                            sameCluster: false,
+                        },
+                    };
+                };
+
+                const wsi = await this.workspaceDb.trace(ctx).findCurrentInstance(workspaceID);
+                if (!wsi || wsi.stoppedTime !== undefined) {
+                    return;
+                }
+
+                // (AT) At this point we found a running/building prebuild, which might also include
+                // image build in current state.
+                //
+                // The owner's client connection is automatically registered to listen on instance updates.
+                // For the remaining client connections which would handle `createWorkspace` and end up here, it
+                // also would be reasonable to listen on the instance updates of a running prebuild, or image build.
+                //
+                // We need to be forwarded the WorkspaceInstanceUpdates in the frontend, because we do not have
+                // any other means to reliably learn about the status about image builds, yet.
+                // Once we have those, we should remove this.
+                //
+                const ws = await this.workspaceDb.trace(ctx).findById(workspaceID);
+                if (!!ws && !!wsi && ws.ownerId !== this.user?.id) {
+                    const resetListener = this.localMessageBroker.listenForWorkspaceInstanceUpdates(
+                        ws.ownerId,
+                        (ctx, instance) => {
+                            if (instance.id === wsi.id) {
+                                this.forwardInstanceUpdateToClient(ctx, instance);
+                                if (instance.status.phase === "stopped") {
+                                    resetListener.dispose();
+                                }
+                            }
+                        },
+                    );
+                    this.disposables.push(resetListener);
+                }
+
+                const result = makeResult(wsi.id);
+
+                const inSameCluster = wsi.region === this.config.installationShortname;
+                if (!inSameCluster) {
+                    /* We need to wait for this prebuild to finish before we return from here.
+                     * This creation mode is meant to be used once we have gone through default mode, have confirmation from the
+                     * message bus that the prebuild is done, and now only have to wait for dbsync to come through. Thus,
+                     * in this mode we'll poll the database until the prebuild is ready (or we time out).
+                     *
+                     * Note: This polling mechanism only makes sense if the prebuild runs in cluster different from ours.
+                     *       Otherwise there's no dbsync inbetween that we might have to wait for.
+                     *
+                     * DB sync interval is 2 seconds at the moment, we wait ten "ticks" for the data to be synchronized.
+                     */
+                    const finishedPrebuiltWorkspace = await this.pollDatabaseUntilPrebuildIsAvailable(
+                        ctx,
+                        prebuiltWorkspace.id,
+                        20000,
+                    );
+                    if (!finishedPrebuiltWorkspace) {
+                        log.warn(
+                            logCtx,
+                            "did not find a finished prebuild in the database despite waiting long enough after msgbus confirmed that the prebuild had finished",
+                            logPayload,
+                        );
+                        return;
+                    } else {
+                        return {
+                            title: context.title,
+                            originalContext: context,
+                            prebuiltWorkspace: finishedPrebuiltWorkspace,
+                        } as PrebuiltWorkspaceContext;
+                    }
+                }
+
+                /* This is the default mode behaviour: we present the running prebuild to the user so that they can see the logs
+                 * or choose to force the creation of a workspace.
+                 */
+                if (wsi.status.phase != "running") {
+                    result.runningWorkspacePrebuild!.starting = "starting";
+                } else {
+                    result.runningWorkspacePrebuild!.starting = "running";
+                }
+                log.info(
+                    logCtx,
+                    `Found prebuilding (starting=${
+                        result.runningWorkspacePrebuild!.starting
+                    }) workspace for ${cloneUrl}:${commitSHAs}`,
+                    logPayload,
+                );
+                return result;
+            }
+        } catch (e) {
+            TraceContext.setError(ctx, e);
+            throw e;
+        } finally {
+            ctx.span.finish();
+        }
     }
 
     protected listenForWorkspaceInstanceUpdates(): void {
@@ -1092,8 +1324,16 @@ export class GitpodServerImpl implements GitpodServerWithTracing, Disposable {
         }
     }
 
-    public async isPrebuildDone(ctx: TraceContext, pwsid: string): Promise<boolean> {
-        throw new ResponseError(ErrorCodes.EE_FEATURE, "Prebuilds are implemented in Gitpod's enterprise edition");
+    public async isPrebuildDone(ctx: TraceContext, pwsId: string): Promise<boolean> {
+        traceAPIParams(ctx, { pwsId });
+
+        const pws = await this.workspaceDb.trace(ctx).findPrebuildByID(pwsId);
+        if (!pws) {
+            // there is no prebuild - that's as good one being done
+            return true;
+        }
+
+        return PrebuiltWorkspace.isDone(pws);
     }
 
     public async createWorkspace(
@@ -1331,6 +1571,146 @@ export class GitpodServerImpl implements GitpodServerWithTracing, Disposable {
         );
     }
 
+    // Projects
+    async getProviderRepositoriesForUser(
+        ctx: TraceContext,
+        params: { provider: string; hints?: object },
+    ): Promise<ProviderRepository[]> {
+        traceAPIParams(ctx, { params });
+
+        const user = this.checkAndBlockUser("getProviderRepositoriesForUser");
+
+        const repositories: ProviderRepository[] = [];
+        const providerHost = params.provider;
+        const provider = (await this.getAuthProviders(ctx)).find((ap) => ap.host === providerHost);
+
+        if (providerHost === "github.com" && this.config.githubApp?.enabled) {
+            repositories.push(...(await this.githubAppSupport.getProviderRepositoriesForUser({ user, ...params })));
+        } else if (provider?.authProviderType === "GitHub") {
+            const hostContext = this.hostContextProvider.get(providerHost);
+            if (hostContext?.services) {
+                repositories.push(
+                    ...(await hostContext.services.repositoryService.getRepositoriesForAutomatedPrebuilds(user)),
+                );
+            }
+        } else if (providerHost === "bitbucket.org" && provider) {
+            repositories.push(...(await this.bitbucketAppSupport.getProviderRepositoriesForUser({ user, provider })));
+        } else if (provider?.authProviderType === "BitbucketServer") {
+            const hostContext = this.hostContextProvider.get(providerHost);
+            if (hostContext?.services) {
+                repositories.push(
+                    ...(await hostContext.services.repositoryService.getRepositoriesForAutomatedPrebuilds(user)),
+                );
+            }
+        } else if (provider?.authProviderType === "GitLab") {
+            repositories.push(...(await this.gitLabAppSupport.getProviderRepositoriesForUser({ user, provider })));
+        } else {
+            log.info({ userId: user.id }, `Unsupported provider: "${params.provider}"`, { params });
+        }
+        const projects = await this.projectsService.getProjectsByCloneUrls(repositories.map((r) => r.cloneUrl));
+
+        const cloneUrlToProject = new Map(projects.map((p) => [p.cloneUrl, p]));
+
+        for (const repo of repositories) {
+            const p = cloneUrlToProject.get(repo.cloneUrl);
+            const repoProvider = new URL(repo.cloneUrl).host.split(".")[0];
+
+            if (p) {
+                if (p.userId) {
+                    const owner = await this.userDB.findUserById(p.userId);
+                    if (owner) {
+                        const ownerProviderMatchingRepoProvider = owner.identities.find((identity, index) =>
+                            identity.authProviderId.toLowerCase().includes(repoProvider),
+                        );
+                        if (ownerProviderMatchingRepoProvider) {
+                            repo.inUse = {
+                                userName: ownerProviderMatchingRepoProvider?.authName,
+                            };
+                        }
+                    }
+                } else if (p.teamOwners && p.teamOwners[0]) {
+                    repo.inUse = {
+                        userName: p.teamOwners[0] || "somebody",
+                    };
+                }
+            }
+        }
+
+        return repositories;
+    }
+
+    public async getPrebuildEvents(ctx: TraceContext, projectId: string): Promise<PrebuildEvent[]> {
+        traceAPIParams(ctx, { projectId });
+        const user = this.checkAndBlockUser("getPrebuildEvents");
+
+        const project = await this.projectsService.getProject(projectId);
+        if (!project) {
+            throw new ResponseError(ErrorCodes.NOT_FOUND, "Project not found");
+        }
+        await this.guardProjectOperation(user, projectId, "get");
+
+        const events = await this.projectsService.getPrebuildEvents(project.cloneUrl);
+        return events;
+    }
+
+    async triggerPrebuild(
+        ctx: TraceContext,
+        projectId: string,
+        branchName: string | null,
+    ): Promise<StartPrebuildResult> {
+        traceAPIParams(ctx, { projectId, branchName });
+
+        const user = this.checkAndBlockUser("triggerPrebuild");
+
+        const project = await this.projectsService.getProject(projectId);
+        if (!project) {
+            throw new ResponseError(ErrorCodes.NOT_FOUND, "Project not found");
+        }
+        await this.guardProjectOperation(user, projectId, "update");
+
+        const branchDetails = !!branchName
+            ? await this.projectsService.getBranchDetails(user, project, branchName)
+            : (await this.projectsService.getBranchDetails(user, project)).filter((b) => b.isDefault);
+        if (branchDetails.length !== 1) {
+            log.debug({ userId: user.id }, "Cannot find branch details.", { project, branchName });
+            throw new ResponseError(
+                ErrorCodes.NOT_FOUND,
+                `Could not find ${!branchName ? "a default branch" : `branch '${branchName}'`} in repository ${
+                    project.cloneUrl
+                }`,
+            );
+        }
+        const contextURL = branchDetails[0].url;
+
+        const context = (await this.contextParser.handle(ctx, user, contextURL)) as CommitContext;
+
+        // HACK: treat manual triggered prebuild as a reset for the inactivity state
+        await this.projectDB.updateProjectUsage(project.id, {
+            lastWorkspaceStart: new Date().toISOString(),
+        });
+
+        const prebuild = await this.prebuildManager.startPrebuild(ctx, {
+            context,
+            user,
+            project,
+            forcePrebuild: true,
+        });
+
+        this.analytics.track({
+            userId: user.id,
+            event: "prebuild_triggered",
+            properties: {
+                context_url: contextURL,
+                clone_url: project.cloneUrl,
+                commit: context.revision,
+                branch: branchDetails[0].name,
+                project_id: project.id,
+            },
+        });
+
+        return prebuild;
+    }
+
     /**
      * This is an explicit extension point for allowing downstream versions of Gitpod to selectively restrict access.
      * @returns true if the user is allowed to access the repository
@@ -1349,17 +1729,6 @@ export class GitpodServerImpl implements GitpodServerWithTracing, Disposable {
                 }
             } catch {}
         }
-        return undefined;
-    }
-
-    protected async findPrebuiltWorkspace(
-        parentCtx: TraceContext,
-        user: User,
-        context: WorkspaceContext,
-        ignoreRunningPrebuild?: boolean,
-        allowUsingPreviousPrebuilds?: boolean,
-    ): Promise<WorkspaceCreationResult | PrebuiltWorkspaceContext | undefined> {
-        // prebuilds are an EE feature
         return undefined;
     }
 
@@ -2611,35 +2980,6 @@ export class GitpodServerImpl implements GitpodServerWithTracing, Disposable {
         }
     }
 
-    public async getProviderRepositoriesForUser(
-        ctx: TraceContext,
-        params: { provider: string },
-    ): Promise<ProviderRepository[]> {
-        traceAPIParams(ctx, { params });
-
-        this.checkAndBlockUser("getProviderRepositoriesForUser");
-        // Note: this operation is per-user only, hence needs no resource guard
-
-        // implemented in EE
-        return [];
-    }
-
-    public async createProject(ctx: TraceContext, params: CreateProjectParams): Promise<Project> {
-        traceAPIParams(ctx, { params });
-
-        const user = this.checkUser("createProject");
-        if (params.userId) {
-            if (params.userId !== user.id) {
-                throw new ResponseError(ErrorCodes.PERMISSION_DENIED, `Cannot create Projects for other users`);
-            }
-        } else {
-            // Anyone who can read a team's information (i.e. any team member) can create a new project.
-            await this.guardTeamOperation(params.teamId || "", "get", "not_implemented");
-        }
-
-        return this.projectsService.createProject(params, user);
-    }
-
     public async deleteProject(ctx: TraceContext, projectId: string): Promise<void> {
         traceAPIParams(ctx, { projectId });
 
@@ -2806,32 +3146,65 @@ export class GitpodServerImpl implements GitpodServerWithTracing, Disposable {
         }
     }
 
-    public async getPrebuildEvents(ctx: TraceContext, projectId: string): Promise<PrebuildEvent[]> {
-        this.checkAndBlockUser("getPrebuildEvents");
-        throw new ResponseError(
-            ErrorCodes.EE_FEATURE,
-            `Prebuild Events are implemented in Gitpod's Enterprise Edition`,
-        );
+    async adminFindPrebuilds(ctx: TraceContext, params: FindPrebuildsParams): Promise<PrebuildWithStatus[]> {
+        traceAPIParams(ctx, { params });
+        await this.guardAdminAccess("adminFindPrebuilds", { params }, Permission.ADMIN_PROJECTS);
+
+        return this.projectsService.findPrebuilds(params);
     }
 
-    public async triggerPrebuild(
-        ctx: TraceContext,
-        projectId: string,
-        branchName: string | null,
-    ): Promise<StartPrebuildResult> {
-        this.checkAndBlockUser("triggerPrebuild");
-        throw new ResponseError(
-            ErrorCodes.EE_FEATURE,
-            `Triggering Prebuilds is implemented in Gitpod's Enterprise Edition`,
-        );
+    async cancelPrebuild(ctx: TraceContext, projectId: string, prebuildId: string): Promise<void> {
+        traceAPIParams(ctx, { projectId, prebuildId });
+
+        const user = this.checkAndBlockUser("cancelPrebuild");
+
+        const project = await this.projectsService.getProject(projectId);
+        if (!project) {
+            throw new ResponseError(ErrorCodes.NOT_FOUND, "Project not found");
+        }
+        await this.guardProjectOperation(user, projectId, "update");
+
+        const prebuild = await this.workspaceDb.trace(ctx).findPrebuildByID(prebuildId);
+        if (!prebuild) {
+            throw new ResponseError(ErrorCodes.NOT_FOUND, "Prebuild not found");
+        }
+        // Explicitly stopping the prebuild workspace now automaticaly cancels the prebuild
+        await this.stopWorkspace(ctx, prebuild.buildWorkspaceId);
     }
 
-    public async cancelPrebuild(ctx: TraceContext, projectId: string, prebuildId: string): Promise<void> {
-        this.checkAndBlockUser("cancelPrebuild");
-        throw new ResponseError(
-            ErrorCodes.EE_FEATURE,
-            `Cancelling Prebuilds is implemented in Gitpod's Enterprise Edition`,
+    public async createProject(ctx: TraceContext, params: CreateProjectParams): Promise<Project> {
+        traceAPIParams(ctx, { params });
+
+        const user = this.checkUser("createProject");
+        if (params.userId) {
+            if (params.userId !== user.id) {
+                throw new ResponseError(ErrorCodes.PERMISSION_DENIED, `Cannot create Projects for other users`);
+            }
+        } else {
+            // Anyone who can read a team's information (i.e. any team member) can create a new project.
+            await this.guardTeamOperation(params.teamId || "", "get", "not_implemented");
+        }
+
+        const project = await this.projectsService.createProject(params, user);
+        // update client registration for the logged in user
+        this.disposables.push(
+            this.localMessageBroker.listenForPrebuildUpdates(
+                project.id,
+                (ctx: TraceContext, update: PrebuildWithStatus) =>
+                    TraceContext.withSpan(
+                        "forwardPrebuildUpdateToClient",
+                        (ctx) => {
+                            traceClientMetadata(ctx, this.clientMetadata);
+                            TraceContext.setJsonRPCMetadata(ctx, "onPrebuildUpdate");
+
+                            this.client?.onPrebuildUpdate(update);
+                        },
+                        ctx,
+                    ),
+            ),
         );
+
+        return project;
     }
 
     public async updateProjectPartial(ctx: TraceContext, partialProject: PartialProject): Promise<void> {
@@ -3315,10 +3688,6 @@ export class GitpodServerImpl implements GitpodServerWithTracing, Disposable {
     ): Promise<void> {
         await this.guardAdminAccess("adminSetTeamMemberRole", { teamId, userId, role }, Permission.ADMIN_WORKSPACES);
         return this.teamDB.setTeamMemberRole(userId, teamId, role);
-    }
-
-    async adminFindPrebuilds(ctx: TraceContext, params: FindPrebuildsParams): Promise<PrebuildWithStatus[]> {
-        throw new ResponseError(ErrorCodes.EE_FEATURE, `Admin support is implemented in Gitpod's Enterprise Edition`);
     }
 
     async adminGetSettings(ctx: TraceContext): Promise<InstallationAdminSettings> {
